sauce it down, start on core logic
