+++ src/strategies/optuna.py
@@ -1,6 +1,4 @@
 import optuna
-## objective function optimization by 3rd party service
+# objective function optimization by 3rd party service
 
 study = optuna.create_study()
-
-
